+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - auth sub folder
 - login schema
 - register schema
+- lost password schema
 
 ### Changed
 
@@ -25,6 +26,8 @@ All notable changes to this project will be documented in this file.
 - convert login form hook to use react hook form
 - convert register form component to use react hook form
 - convert register form hook to use react hook form
+- convert lost password form component to use react hook form
+- convert lost password form hook to use react hook form
 
 ## [0.1.1] - 2026-01-14